created custom hook for pagination 1 attempt
