+++ src/pages/order-history/order-history.js
@@ -1,7 +1,5 @@
-import React, { useState } from 'react';
-import { useQuery } from '@apollo/client';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { Loader } from '../../components/loader/loader';
 import OrderHistoryOrder from '../../containers/orders/order-history/order-history-order';
 import EmptyOrderHistory from '../../containers/orders/order-history/empty-order-history';
 import OrderHistoryPagination from '../../containers/orders/order-history/order-history-pagination/index';
@@ -9,40 +7,13 @@ import { useStyles } from './order-history.styles';
 import { limitHistoryOrders } from '../../const/user-order-history';
 import errorOrLoadingHandler from '../../utils/errorOrLoadingHandler';
 import { getUserOrdersQuery } from './operations/order-history.queries';
+import usePaginationForOrders from './hooks/use-pagination-for-orders';
 
 const OrderHistory = () => {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [ordersCount, setOrdersCount] = useState(null);
-  const [orders, setOrders] = useState([]);
+  const { loadingOrders, errorOrders, orders, currentPage, changeHandler, quantityPages } =
+    usePaginationForOrders(limitHistoryOrders, getUserOrdersQuery);
   const { t } = useTranslation();
   const styles = useStyles();
-
-  const { loading: loadingOrders, error: errorOrders } = useQuery(getUserOrdersQuery, {
-    variables: {
-      pagination: {
-        limit: limitHistoryOrders,
-        skip: (currentPage - 1) * limitHistoryOrders
-      }
-    },
-    onCompleted: (data) => {
-      setOrders(data.getUserOrders.userOrders);
-      setOrdersCount(data.getUserOrders.ordersCount);
-    }
-  });
-
-  const quantityPages = Math.ceil(ordersCount / limitHistoryOrders);
-
-  const changeHandler = (value) => {
-    setCurrentPage(value);
-  };
-
-  if (loadingOrders) {
-    return (
-      <div className={styles.loader}>
-        <Loader className={styles.loader} />
-      </div>
-    );
-  }
 
   if (loadingOrders || errorOrders) return errorOrLoadingHandler(errorOrders, loadingOrders);
